refactor(stellar): split soroswap.ts (1078 lines) into focused modules
- soroswap/tokens.ts: token registry, getTokens(), getAvailableTokens(), getTokenAddress()
- soroswap/utils.ts: SDK management, API key helpers, amount conversion, formatting
- soroswap/quotes.ts: getQuote(), getPool(), buildTransaction(), sendTransaction(), addLiquidity()
- soroswap/index.ts: barrel re-exporting all public API
- soroswap.ts: thin re-export shim for backwards compatibility (no import changes needed)
