+++ apps/web-app/src/lib/helpers/soroswap/index.ts
@@ -1,0 +1,51 @@
+/**
+ * Soroswap helpers barrel
+ * Re-exports all public API for backwards compatibility.
+ *
+ * Internal modules:
+ *   tokens.ts  — token registry and lookup
+ *   utils.ts   — SDK management, API key, formatting helpers
+ *   quotes.ts  — quote, pool, build, send, and liquidity functions
+ */
+
+// Types
+export type {
+  Token,
+  QuoteRequest,
+  QuoteResponse,
+  BuildRequest,
+  BuildResponse,
+  SendRequest,
+  SendResponse,
+  AddLiquidityRequest,
+  AddLiquidityResponse,
+  PoolInfo,
+  GetPoolRequest,
+} from "../../types/soroswapTypes";
+
+// Token registry
+export {
+  getAvailableTokens,
+  getTokens,
+  getTokenAddress,
+  TOKENS,
+} from "./tokens";
+
+// Utilities
+export {
+  getApiKey,
+  setApiKey,
+  hasApiKey,
+  invalidateSoroswapSDK,
+  isValidContractAddress,
+  getTokenExplorerUrl,
+} from "./utils";
+
+// Quotes, pools, transactions, and liquidity
+export {
+  getPool,
+  getQuote,
+  buildTransaction,
+  sendTransaction,
+  addLiquidity,
+} from "./quotes";
